refactor extractor classes for all types except api
